requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.1
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
iniconfig==2.0.0
packaging==23.2
Pillow==10.0.1
pluggy==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3